Encapsulate kriging equations in a dedicated class (bis).
* core/stk_predict.m: Move the rhs loop (and a few other things)
 to @stk_kriging_equation.stk_predict(). Add tests.
* core/@stk_kriging_equation/stk_predict.m: New function.
* core/@stk_kriging_equation/linsolve.m: Cast xt to double.
* core/@stk_kriging_equation/stk_kriging_equation.m: Cast xi to
 double. Add a third optional argument: 'xt'.

+++ core/@stk_kriging_equation/linsolve.m
@@ -31,7 +31,7 @@ function kreq = linsolve(kreq, xt)
 
 [Kti, Pt] = stk_make_matcov(kreq.model, xt, kreq.xi);
 
-kreq.xt = xt;
+kreq.xt = double(xt);
 kreq.RS = [Kti Pt]';
 
 % solve the upper-triangular system to get the extended
